+++ src/Symfony/Component/Messenger/Tests/Transport/Sender/SendersLocatorTest.php
@@ -14,7 +14,7 @@ namespace Symfony\Component\Messenger\Tests\Transport\Sender;
 use PHPUnit\Framework\TestCase;
 use Psr\Container\ContainerInterface;
 use Symfony\Component\Messenger\Envelope;
-use Symfony\Component\Messenger\Stamp\ViaSenderStamp;
+use Symfony\Component\Messenger\Stamp\TransportNamesStamp;
 use Symfony\Component\Messenger\Tests\Fixtures\DummyMessage;
 use Symfony\Component\Messenger\Tests\Fixtures\SecondMessage;
 use Symfony\Component\Messenger\Transport\Sender\SenderInterface;
@@ -36,7 +36,7 @@ class SendersLocatorTest extends TestCase
         $this->assertSame([], iterator_to_array($locator->getSenders(new Envelope(new SecondMessage()))));
     }
 
-    public function testItReturnsTheSenderBasedOnViaSenderStamp()
+    public function testItReturnsTheSenderBasedOnTransportNamesStamp()
     {
         $mySender = $this->createMock(SenderInterface::class);
         $otherSender = $this->createMock(SenderInterface::class);
@@ -48,7 +48,7 @@ class SendersLocatorTest extends TestCase
             DummyMessage::class => ['my_sender'],
         ], $sendersLocator);
 
-        $this->assertSame(['other_sender' => $otherSender], iterator_to_array($locator->getSenders(new Envelope(new DummyMessage('a'), [new ViaSenderStamp(['other_sender'])]))));
+        $this->assertSame(['other_sender' => $otherSender], iterator_to_array($locator->getSenders(new Envelope(new DummyMessage('a'), [new TransportNamesStamp(['other_sender'])]))));
         $this->assertSame([], iterator_to_array($locator->getSenders(new Envelope(new SecondMessage()))));
     }